requirements.txt
feedparser
requests
python-dotenv
openai>=1.12.0
httpx<0.28.0
alpaca-trade-api
pandas
numpy
streamlit
google-genai
pandas_market_calendars
pytz
